+++ lib/app/services/event/events/event_bookmark_event.dart
@@ -1,0 +1,3 @@
+import 'package:duty_it/app/services/event/events/app_event.dart';
+
+class EventBookmarkEvent extends AppEvent {}
